requirements.txt
appnope==0.1.0
backports.shutil-get-terminal-size==1.0.0
bitly-api==0.3
decorator==4.0.10
eventbrite==3.3.3
ipython==5.0.0
ipython-genutils==0.1.0
numpy==1.11.1
pathlib2==2.1.0
pexpect==4.0.1
pickleshare==0.7.3
prompt-toolkit==1.0.3
ptyprocess==0.5.1
Pygments==2.1.3
python-dateutil==2.4.1
requests==2.10.0
sendgrid==3.6.0
simplegeneric==0.8.1
six==1.10.0
traitlets==4.2.2
wcwidth==0.1.7
